Refactor DerivaML using mixins to reduce class size
Extract common functionality into reusable mixins:
- PathBuilderMixin: pathBuilder(), domain_path, table_path(),
  get_table_as_dataframe(), get_table_as_dict()
- RidResolutionMixin: resolve_rid(), retrieve_rid()
- VocabularyMixin: add_term(), lookup_term(), list_vocabulary_terms()
- WorkflowMixin: list_workflows(), add_workflow(), lookup_workflow(),
  create_workflow()

This reduces DerivaML class size by ~400 lines while improving code
organization and enabling potential reuse by other classes.
