+++ src/deriva_ml/core/mixins/__init__.py
@@ -1,0 +1,24 @@
+"""Mixins for DerivaML catalog operations.
+
+This module provides mixins that can be used to compose catalog-related
+functionality. Each mixin provides a specific set of operations that can
+be mixed into classes that have access to a catalog.
+
+Mixins:
+    VocabularyMixin: Vocabulary term management (add, lookup, list terms)
+    RidResolutionMixin: RID resolution and retrieval
+    PathBuilderMixin: Path building and table access utilities
+    WorkflowMixin: Workflow management (add, lookup, list, create)
+"""
+
+from deriva_ml.core.mixins.vocabulary import VocabularyMixin
+from deriva_ml.core.mixins.rid_resolution import RidResolutionMixin
+from deriva_ml.core.mixins.path_builder import PathBuilderMixin
+from deriva_ml.core.mixins.workflow import WorkflowMixin
+
+__all__ = [
+    "VocabularyMixin",
+    "RidResolutionMixin",
+    "PathBuilderMixin",
+    "WorkflowMixin",
+]
